Refactor I2CBusManager to use modern i2c_master.h driver
- Migrated ALC::I2CBusManager from legacy driver/i2c.h to modern driver/i2c_master.h.
- Implemented internal i2c_master_dev_handle_t caching for efficient address-based communication.
- Updated all I2C address parameters to uint16_t to support both 7-bit and 10-bit devices.
- Refactored BME280Sensor, BNO086Sensor, and MaxM10sSensor to use the new address type and headers.
- Enabled default glitch filtering and internal pull-ups for improved bus stability.
- Ensured all I2C operations remain non-blocking and serialized through the manager task.
- Cleaned up all legacy i2c.h references across the component.

+++ include/alc_bme280_sensor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "driver/i2c.h"
+#include "driver/i2c_master.h"
 #include "esp_err.h"
 #include "alc_i2c_bus_manager.h"
 #include <mutex>
@@ -115,7 +115,7 @@ public:
    * @param bus_manager Reference to the I2C bus manager.
    * @param address I2C address of the sensor (default: 0x76).
    */
-  explicit BME280Sensor(I2CBusManager& bus_manager, uint8_t address = 0x76);
+  explicit BME280Sensor(I2CBusManager& bus_manager, uint16_t address = 0x76);
 
   /**
    * @brief Copying or moving a sensor instance is not allowed.
@@ -158,7 +158,7 @@ private:
   esp_err_t ReadAndProcessData(BusToken& token);
 
   I2CBusManager& bus_manager_;
-  uint8_t address_;
+  uint16_t address_;
   Configuration config_;
 
   struct {
